refactor: changed use permissions name convention
